add stored procedures

+++ hadoop-hdfs-project/hadoop-hdfs-db/src/main/java/org/apache/hadoop/hdfs/db/DatabaseINode2Block.java
@@ -1,8 +1,8 @@
 package org.apache.hadoop.hdfs.db;
 
+import java.sql.CallableStatement;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
-import java.sql.CallableStatement;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
@@ -12,6 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.voltdb.*;
 
 public class DatabaseINode2Block {
   static final Logger LOG = LoggerFactory.getLogger(DatabaseINode2Block.class);
@@ -21,7 +22,8 @@ public class DatabaseINode2Block {
       LOG.debug("INode2Block [insert]: (" + id + "," + blockId + "," + idx + ")");
     }
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "INSERT INTO inode2block(id, blockId, idx) VALUES (?, ?, ?);";
       PreparedStatement pst = conn.prepareStatement(sql);
 
@@ -31,6 +33,7 @@ public class DatabaseINode2Block {
 
       pst.executeUpdate();
       pst.close();
+      Database.getInstance().retConnection(obj);
     } catch (SQLException ex) {
       System.err.println(ex.getMessage());
     }
@@ -45,7 +48,8 @@ public class DatabaseINode2Block {
       String env = System.getenv("DATABASE");
       if (env.equals("VOLT")) {
         // call a stored procedure
-        Connection conn = Database.getInstance().getConnection();
+        DatabaseConnection obj = Database.getInstance().getConnection();
+        Connection conn = obj.getConnection();
         CallableStatement proc = conn.prepareCall("{call InsertINode2Block(?, ?, ?)}");
 
         proc.setLong(1, id);
@@ -53,7 +57,7 @@ public class DatabaseINode2Block {
         List<Integer> idxs = new ArrayList<Integer>();
         for (int i = 0; i < blockIds.size(); ++i) {
           idxs.add(index + i);
-        } 
+        }
         proc.setArray(3, conn.createArrayOf("INT", idxs.toArray(new Integer[blockIds.size()])));
 
         ResultSet rs = proc.executeQuery();
@@ -64,6 +68,7 @@ public class DatabaseINode2Block {
         }
         rs.close();
         proc.close();
+        Database.getInstance().retConnection(obj);
       } else {
         int idx = index;
         int size = blockIds.size();
@@ -72,19 +77,21 @@ public class DatabaseINode2Block {
           idx += 1;
           sql +=
               "("
-                + String.valueOf(id)
-                + ","
-                + String.valueOf(blockIds.get(i))
-                + ","
-                + String.valueOf(idx)
-                + "),";
+                  + String.valueOf(id)
+                  + ","
+                  + String.valueOf(blockIds.get(i))
+                  + ","
+                  + String.valueOf(idx)
+                  + "),";
         }
         sql = sql.substring(0, sql.length() - 1) + ";";
 
-        Connection conn = Database.getInstance().getConnection();
+        DatabaseConnection obj = Database.getInstance().getConnection();
+        Connection conn = obj.getConnection();
         Statement st = conn.createStatement();
         st.executeUpdate(sql);
         st.close();
+        Database.getInstance().retConnection(obj);
         if (LOG.isDebugEnabled()) {
           LOG.debug("INode2Block [insert]: (" + sql + ")");
         }
@@ -96,7 +103,8 @@ public class DatabaseINode2Block {
 
   private static <T> void setAttribute(final long id, final String attrName, final T attrValue) {
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
 
       String sql = "UPDATE inode2block SET " + attrName + " = ? WHERE blockId = ?;";
       PreparedStatement pst = conn.prepareStatement(sql);
@@ -111,12 +119,13 @@ public class DatabaseINode2Block {
         pst.setLong(1, ((Long) attrValue).longValue());
       } else {
         System.err.println("Only support string and long types for now.");
-        System.exit(0);
+        System.exit(-1);
       }
       pst.setLong(2, id);
 
       pst.executeUpdate();
       pst.close();
+      Database.getInstance().retConnection(obj);
     } catch (SQLException ex) {
       System.err.println(ex.getMessage());
     }
@@ -128,7 +137,8 @@ public class DatabaseINode2Block {
   private static <T> T getAttribute(final long id, final String attrName) {
     T result = null;
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "SELECT " + attrName + " FROM inode2block WHERE blockId = ?;";
       PreparedStatement pst = conn.prepareStatement(sql);
       pst.setLong(1, id);
@@ -143,6 +153,7 @@ public class DatabaseINode2Block {
       }
       rs.close();
       pst.close();
+      Database.getInstance().retConnection(obj);
     } catch (SQLException ex) {
       System.err.println(ex.getMessage());
     }
@@ -156,16 +167,32 @@ public class DatabaseINode2Block {
   public static int getNumBlocks(final long id) {
     int num = 0;
     try {
-      Connection conn = Database.getInstance().getConnection();
-      String sql = "SELECT COUNT(DISTINCT blockId) FROM inode2block WHERE id = ?;";
-      PreparedStatement pst = conn.prepareStatement(sql);
-      pst.setLong(1, id);
-      ResultSet rs = pst.executeQuery();
-      while (rs.next()) {
-        num = rs.getInt(1);
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      String env = System.getenv("DATABASE");
+      if (env.equals("VOLT")) {
+        try {
+          VoltTable[] results = obj.getVoltClient().callProcedure("GetNumBlocks", id).getResults();
+          VoltTable result = results[0];
+          result.resetRowPosition();
+          while (result.advanceRow()) {
+            num = (int) result.getLong(0);
+          }
+        } catch (Exception e) {
+          e.printStackTrace();
+        }
+      } else {
+        Connection conn = obj.getConnection();
+        String sql = "SELECT COUNT(DISTINCT blockId) FROM inode2block WHERE id = ?;";
+        PreparedStatement pst = conn.prepareStatement(sql);
+        pst.setLong(1, id);
+        ResultSet rs = pst.executeQuery();
+        while (rs.next()) {
+          num = rs.getInt(1);
+        }
+        rs.close();
+        pst.close();
       }
-      rs.close();
-      pst.close();
+      Database.getInstance().retConnection(obj);
     } catch (SQLException ex) {
       System.out.println(ex.getMessage());
     }
@@ -180,7 +207,8 @@ public class DatabaseINode2Block {
   public static int getLastBlockId(final long id) {
     int blockId = -1;
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "SELECT blockId FROM inode2block WHERE id = ? ORDER BY idx DESC LIMIT 1;";
       PreparedStatement pst = conn.prepareStatement(sql);
       pst.setLong(1, id);
@@ -190,6 +218,7 @@ public class DatabaseINode2Block {
       }
       rs.close();
       pst.close();
+      Database.getInstance().retConnection(obj);
     } catch (SQLException ex) {
       System.out.println(ex.getMessage());
     }
@@ -204,7 +233,8 @@ public class DatabaseINode2Block {
   public static long getBcId(final long blockId) {
     long id = 0;
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "SELECT id FROM inode2block WHERE blockId = ?;";
       PreparedStatement pst = conn.prepareStatement(sql);
       pst.setLong(1, blockId);
@@ -214,6 +244,7 @@ public class DatabaseINode2Block {
       }
       rs.close();
       pst.close();
+      Database.getInstance().retConnection(obj);
       if (LOG.isDebugEnabled()) {
         LOG.debug("getBcId: (" + blockId + "," + id + ")");
       }
@@ -227,7 +258,8 @@ public class DatabaseINode2Block {
   public static long getSize() {
     long size = 0;
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "SELECT COUNT(blockId) FROM inode2block;";
       Statement st = conn.createStatement();
       ResultSet rs = st.executeQuery(sql);
@@ -236,6 +268,7 @@ public class DatabaseINode2Block {
       }
       rs.close();
       st.close();
+      Database.getInstance().retConnection(obj);
       if (LOG.isDebugEnabled()) {
         LOG.debug("getSize: (" + size + ")");
       }
@@ -243,7 +276,7 @@ public class DatabaseINode2Block {
       System.out.println(ex.getMessage());
     }
 
-    return size;    
+    return size;
   }
 
   public static void setBcIdViaBlkId(final long blockId, final long bcId) {
@@ -252,13 +285,15 @@ public class DatabaseINode2Block {
 
   public static void setBcIdViaBcId(final long bcId, final long newBcId) {
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "UPDATE inode2block SET id = ? WHERE id = ?;";
       PreparedStatement pst = conn.prepareStatement(sql);
       pst.setLong(1, newBcId);
       pst.setLong(2, bcId);
       pst.executeUpdate();
       pst.close();
+      Database.getInstance().retConnection(obj);
       if (LOG.isDebugEnabled()) {
         LOG.debug("setBcIdViaBcId: (" + bcId + "," + newBcId + "," + sql + ")");
       }
@@ -270,7 +305,8 @@ public class DatabaseINode2Block {
   public static List<Long> getBlockIds(final long inodeId) {
     List<Long> blockIds = new ArrayList<>();
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "SELECT blockId FROM inode2block WHERE id = ?;";
       PreparedStatement pst = conn.prepareStatement(sql);
       pst.setLong(1, inodeId);
@@ -281,6 +317,7 @@ public class DatabaseINode2Block {
       }
       rs.close();
       pst.close();
+      Database.getInstance().retConnection(obj);
     } catch (SQLException ex) {
       System.out.println(ex.getMessage());
     }
@@ -293,7 +330,8 @@ public class DatabaseINode2Block {
   public static List<Long> getAllBlockIds() {
     List<Long> blockIds = new ArrayList<>();
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "SELECT blockId FROM inode2block;";
       Statement st = conn.createStatement();
       ResultSet rs = st.executeQuery(sql);
@@ -303,6 +341,7 @@ public class DatabaseINode2Block {
       }
       rs.close();
       st.close();
+      Database.getInstance().retConnection(obj);
     } catch (SQLException ex) {
       System.out.println(ex.getMessage());
     }
@@ -314,12 +353,14 @@ public class DatabaseINode2Block {
 
   public static void deleteViaBlkId(final long blockId) {
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "DELETE FROM inode2block WHERE blockId = ?;";
       PreparedStatement pst = conn.prepareStatement(sql);
       pst.setLong(1, blockId);
       pst.executeUpdate();
       pst.close();
+      Database.getInstance().retConnection(obj);
       if (LOG.isDebugEnabled()) {
         LOG.debug("deleteViaBlkId: (" + blockId + "," + sql + ")");
       }
@@ -330,13 +371,15 @@ public class DatabaseINode2Block {
 
   public static void delete(final long nodeId, final int idx) {
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "DELETE FROM inode2block WHERE id = ? and idx = ?;";
       PreparedStatement pst = conn.prepareStatement(sql);
       pst.setLong(1, nodeId);
       pst.setInt(2, idx);
       pst.executeUpdate();
       pst.close();
+      Database.getInstance().retConnection(obj);
       if (LOG.isDebugEnabled()) {
         LOG.debug("delete: (" + nodeId + "," + idx + "," + sql + ")");
       }
@@ -347,14 +390,25 @@ public class DatabaseINode2Block {
 
   public static void deleteViaBcId(final long nodeId) {
     try {
-      Connection conn = Database.getInstance().getConnection();
-      String sql = "DELETE FROM inode2block WHERE id = ?;";
-      PreparedStatement pst = conn.prepareStatement(sql);
-      pst.setLong(1, nodeId);
-      pst.executeUpdate();
-      pst.close();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      String env = System.getenv("DATABASE");
+      if (env.equals("VOLT")) {
+        try {
+          obj.getVoltClient().callProcedure("DeleteViaBcId", nodeId);
+        } catch (Exception e) {
+          e.printStackTrace();
+        }
+      } else {
+        Connection conn = obj.getConnection();
+        String sql = "DELETE FROM inode2block WHERE id = ?;";
+        PreparedStatement pst = conn.prepareStatement(sql);
+        pst.setLong(1, nodeId);
+        pst.executeUpdate();
+        pst.close();
+      }
+      Database.getInstance().retConnection(obj);
       if (LOG.isDebugEnabled()) {
-        LOG.debug("deleteViaBcId: (" + nodeId + "," + sql + ")");
+        LOG.debug("deleteViaBcId: (" + nodeId + ")");
       }
     } catch (SQLException ex) {
       System.err.println(ex.getMessage());
@@ -363,13 +417,15 @@ public class DatabaseINode2Block {
 
   public static void truncate(final long nodeId, final int n) {
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "DELETE FROM inode2block WHERE id = ? and idx >= ?;";
       PreparedStatement pst = conn.prepareStatement(sql);
       pst.setLong(1, nodeId);
       pst.setInt(2, n);
       pst.executeUpdate();
       pst.close();
+      Database.getInstance().retConnection(obj);
       if (LOG.isDebugEnabled()) {
         LOG.debug("truncate: (" + nodeId + "," + n + "," + sql + ")");
       }
@@ -380,7 +436,8 @@ public class DatabaseINode2Block {
 
   public static void setBlockId(final long nodeId, final int idx, final long blockId) {
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "UPDATE inode2block SET blockId = ? WHERE id = ? and idx = ?;";
       PreparedStatement pst = conn.prepareStatement(sql);
       pst.setLong(1, blockId);
@@ -388,6 +445,7 @@ public class DatabaseINode2Block {
       pst.setInt(3, idx);
       pst.executeUpdate();
       pst.close();
+      Database.getInstance().retConnection(obj);
       if (LOG.isDebugEnabled()) {
         LOG.debug("setBlockId: (" + nodeId + "," + blockId + "," + idx + ")");
       }
@@ -399,7 +457,8 @@ public class DatabaseINode2Block {
   public static int getBlockId(final long nodeId, final int idx) {
     int blockId = -1;
     try {
-      Connection conn = Database.getInstance().getConnection();
+      DatabaseConnection obj = Database.getInstance().getConnection();
+      Connection conn = obj.getConnection();
       String sql = "SELECT blockId from inode2block WHERE id = ? and idx = ?;";
       PreparedStatement pst = conn.prepareStatement(sql);
       pst.setLong(1, nodeId);
@@ -410,6 +469,7 @@ public class DatabaseINode2Block {
       }
       rs.close();
       pst.close();
+      Database.getInstance().retConnection(obj);
       if (LOG.isDebugEnabled()) {
         LOG.debug("getBlockId: (" + nodeId + "," + blockId + ")");
       }
